[FileUploader] Rename location by uri on File model
