+++ src/domain/files/file.model.ts
@@ -3,20 +3,15 @@ export class File {
 
   name: string;
 
-  location: string;
+  uri: string;
 
   size: number;
 
   // noinspection DuplicatedCode
-  public constructor(
-    _id: string,
-    name: string,
-    location: string,
-    size: number,
-  ) {
+  public constructor(_id: string, name: string, uri: string, size: number) {
     this._id = _id;
     this.name = name;
-    this.location = location;
+    this.uri = uri;
     this.size = size;
   }
 }
