+++ src/domain/users/services/errors/avatar-manager-unknown-user.error.ts
