+++ src/domain/users/services/errors/avatar-manager-uploaded-file.error.ts
