+++ src/domain/users/services/errors/password-updater-unknown-user.error.ts
