+++ src/domain/users/services/errors/profile-displayer-unknown-user.error.ts
