+++ src/domain/users/services/errors/profile-updater-unknown-user.error.ts
