+++ src/domain/users/services/errors/register-duplicated-email.error.ts
