+++ src/domain/users/services/errors/register-send-email.error.ts
