+++ src/domain/users/services/errors/suspender-unknown-user.error.ts
