+++ src/infrastructure/emails/sendin-blue-sender.impl.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger } from '@nestjs/common';
 import * as SibApiV3Sdk from 'sib-api-v3-sdk';
 import { IEmailsSender } from './emails-sender.interface';
-import { SendinBlueApiError } from './sendin.blue.api.error';
+import { SendinBlueApiError } from './sendin-blue-api.error';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
