+++ src/infrastructure/mappers/file.mapper.ts
@@ -7,16 +7,16 @@ export class FileMapper {
     return new File(
       fileEntity._id,
       fileEntity.name,
-      fileEntity.location,
+      fileEntity.uri,
       fileEntity.size,
     );
   }
 
   public static toPersistence(file: File): FileEntity {
-    return new FileEntity(file._id, file.name, file.location, file.size);
+    return new FileEntity(file._id, file.name, file.uri, file.size);
   }
 
   public static toDTO(file: File): FileDTO {
-    return new FileDTO(file._id, file.name, file.location, file.size);
+    return new FileDTO(file._id, file.name, file.uri, file.size);
   }
 }
